pyproject.toml
[build-system]
requires = [ "poetry-core",]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "sigfig"
version = "1.3.3"
description = "Python library for rounding numbers (with expected results)"
authors = [ "Michael Busuttil <mike.busuttil@gmail.com>", "Travis Valdez <valdezt@gmail.com>",]

[tool.poetry.dependencies]
python = "^3.11"
sortedcontainers = "^2.4.0"

[tool.poetry.group.dev.dependencies]
semver = "^3.0.2"
pyyaml = "^6.0.1"
requests = "^2.31.0"
toml = "^0.10.2"
numpy = "^1.26.2"
coverage = "^7.3.2"
